(chore): update deps and split tests
