+++ starlite/middleware/compression.py
@@ -49,18 +49,18 @@ class CompressionFacade:
                 from brotli import MODE_FONT, MODE_GENERIC, MODE_TEXT, Compressor
             except ImportError as e:
                 raise MissingDependencyException("brotli is not installed") from e
-            else:
-                modes: Dict[Literal["generic", "text", "font"], int] = {
-                    "text": int(MODE_TEXT),
-                    "font": int(MODE_FONT),
-                    "generic": int(MODE_GENERIC),
-                }
-                self.compressor = Compressor(
-                    quality=config.brotli_quality,
-                    mode=modes[config.brotli_mode],
-                    lgwin=config.brotli_lgwin,
-                    lgblock=config.brotli_lgblock,
-                )
+
+            modes: Dict[Literal["generic", "text", "font"], int] = {
+                "text": int(MODE_TEXT),
+                "font": int(MODE_FONT),
+                "generic": int(MODE_GENERIC),
+            }
+            self.compressor = Compressor(
+                quality=config.brotli_quality,
+                mode=modes[config.brotli_mode],
+                lgwin=config.brotli_lgwin,
+                lgblock=config.brotli_lgblock,
+            )
         else:
             self.compressor = GzipFile(mode="wb", fileobj=buffer, compresslevel=config.gzip_compress_level)
